+++ src/main/java/br/com/inatel/MyCatalog/model/form/ShowForm.java
@@ -17,6 +17,6 @@ public class ShowForm {
     @NotEmpty
     private String title;
     @Min(0) @Max(10)
-    private int personalScore;
+    private double personalScore;
 
 }
